Rename machine client helper module
